Drop jest in favor of cypress components testing
