--- packages/components/.babelrc
@@ -1,7 +1,0 @@
-{
-  "presets": [
-    ["@babel/preset-env", { "targets": { "node": "current" } }],
-    "@babel/preset-typescript",
-    ["@babel/preset-react", { "runtime": "automatic" }]
-  ]
-}
