+++ packages/components/cypress/support/commands.ts
@@ -1,13 +1,24 @@
-Cypress.Commands.add('visitStory', (story, variant) => {
+export {};
+
+Cypress.Commands.add('visitStory', visitStory);
+function visitStory(
+  story: string,
+  variant: string,
+): Cypress.Chainable<Cypress.AUTWindow> {
   let base = Cypress.config('baseUrl') ?? '';
   let url = new URL('/iframe.html', base);
   url.searchParams.set('id', `${story}--${variant}`);
   url.searchParams.set('viewMode', `story`);
 
   return cy.visit(url.toString());
-});
+}
 
-Cypress.Commands.add('compareRects', (target1, target2, compare) => {
+Cypress.Commands.add('compareRects', compareRects);
+function compareRects(
+  target1: () => Cypress.Chainable<JQuery<HTMLElement>>,
+  target2: () => Cypress.Chainable<JQuery<HTMLElement>>,
+  compare: (rect1: DOMRect, rect2: DOMRect) => void,
+) {
   target1().then((target1El) => {
     let target1Rect = target1El[0].getBoundingClientRect();
     target2().then((target2El) => {
@@ -15,12 +26,23 @@ Cypress.Commands.add('compareRects', (target1, target2, compare) => {
       compare(target1Rect, target2Rect);
     });
   });
-});
+}
 
-Cypress.Commands.add('requestAnimationFrame', () => {
+Cypress.Commands.add('requestAnimationFrame', requestAnimationFrame);
+function requestAnimationFrame() {
   return cy.wrap(
     new Promise((resolve) => {
       window.requestAnimationFrame(resolve);
     }),
   );
-});
+}
+
+declare global {
+  namespace Cypress {
+    interface Chainable {
+      visitStory: typeof visitStory;
+      compareRects: typeof compareRects;
+      requestAnimationFrame: typeof requestAnimationFrame;
+    }
+  }
+}
