--- packages/components/jest.config.js
@@ -1,5 +1,0 @@
-const { createJestConfig } = require('@fransvilhelm/config');
-
-module.exports = createJestConfig('jsdom', {
-  coverageDirectory: 'coverage/jest',
-});
