+++ packages/components/package.json
@@ -25,12 +25,9 @@
   "scripts": {
     "dev": "build-package --watch",
     "build": "build-package",
-    "test": "jest",
-    "test:ci": "jest --coverage",
-    "test:ci:react-18": "REACT_VERSION=18 jest",
-    "e2e": "cypress run-ct",
-    "e2e:watch": "cypress open-ct",
-    "lint": "eslint . --ext .ts --ext .tsx --ext .js --ignore-pattern dist",
+    "test": "cypress open-cy",
+    "test:ci": "cypress run-ct",
+    "lint": "eslint src cypress --ext .ts --ext .tsx --ext .js --ignore-pattern dist",
     "storybook": "start-storybook -p 6006",
     "storybook:build": "build-storybook"
   },
@@ -53,10 +50,6 @@
     "@storybook/addon-links": "^6.4.9",
     "@storybook/react": "^6.4.9",
     "@testing-library/cypress": "^8.0.2",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^13.0.0-alpha.5",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.0.3",
     "@types/lodash.debounce": "^4.0.6",
     "@types/node": "^16.11.13",
     "@types/react": "^17.0.37",
@@ -66,12 +59,8 @@
     "babel-plugin-istanbul": "^6.1.1",
     "cypress": "^9.1.1",
     "html-webpack-plugin": "^5.5.0",
-    "jest": "^27.4.5",
     "react": "17.0.2",
-    "react-18": "npm:react@^18.0.0-beta-c7917fe76-20211206",
     "react-dom": "17.0.2",
-    "react-dom-18": "npm:react-dom@^18.0.0-beta-c7917fe76-20211206",
-    "start-server-and-test": "^1.14.0",
     "typescript": "^4.5.4",
     "webpack": "^5.65.0",
     "webpack-dev-server": "^4.6.0"
